requirements.txt
streamlit
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-core
google-generativeai
